Reorganize PEtab library
