+++ petab/__init__.py
@@ -1,7 +1,12 @@
 """PEtab exports"""
 
+from .conditions import *  # noqa: F403, F401
 from .core import *  # noqa: F403, F401
 from .lint import *  # noqa: F403, F401
-from .sbml import *  # noqa: F403, F401
+from .measurements import *  # noqa: F403, F401
 from .parameter_mapping import *  # noqa: F403, F401
+from .parameters import *  # noqa: F403, F401
+from .problem import *  # noqa: F403, F401
+from .sampling import *  # noqa: F403, F401
+from .sbml import *  # noqa: F403, F401
 from .version import __version__  # noqa: F401
